+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 2.0.0
+
+* Added option to show mouse pointer:
+```dart
+onPointer: true, // or false
+```
+
+* Added option to save images to cache (Not available for Web):
+```dart
+cacheImageAndroidIos: true, // or false
+```
+
+* Added option to change the fit of android and ios image. You can define it like this:
+```dart
+fitCarouselList: BoxFit.cover
+```
+
+* Added function to get click on the image:
+```dart
+onTap: () {
+    debugPrint("©gabriel_patrick_souza");
+},
+```
+
 ## 1.0.0
 
 * Initial release.